requirements.txt
fastapi==0.115.12
numpy==2.2.6
protobuf==6.31.0
python-dotenv==1.1.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
uvicorn
google-generativeai